dp_matrix_factorization/requirements.txt
tensorflow_federated~=0.24.0
jax~=0.2.27
tensorflow-privacy~=0.8.0
tensorflow~=2.8.0
pandas~=1.1.5
numpy~=1.21.4
absl-py~=1.0.0
scipy~=1.5.0
